Add AsyncStrategy
