+++ lib/snowagent/agent.rb
@@ -1,20 +1,13 @@
 module SnowAgent
   class Agent
     def initialize(configuration)
-      @queue = Queue.new
-
-      run_sender(configuration)
-    end
-
-    def run_sender(conf)
-      sender_thread = Thread.new { Sender.new(@queue, conf).run }
-      sender_thread.abort_on_exception = true
+      @strategy = AsyncStrategy.new(configuration)
     end
 
     Metric = Struct.new(:name, :value, :at)
 
     def metric(key, value, time = Time.now.to_i)
-      @queue.push(Metric.new(key, value, time))
+      @strategy.metric(Metric.new(key, value, time))
     end
   end
 end
